refactor: colocate useMpdGoalRows with its only consumers
Hook is now fully coupled to GoalCalculatorContext; move from src/hooks to SummaryReport/MpdGoal per hook-placement convention
